Removed raise on missing sendgrid key at compile time
With the raise during compile time, docker builds that are smartly caching the mix dependecy steps
will fail due to not having a config directory available yet. This also should not be tripping at
compile time, but preferably at launch when in MIX_ENV=prod. Will work on addressing that later.